Source/Python/requirements.txt
Flask
gTTS
openai-whisper
ffmpeg-python
requests
python-dotenv